Consolidate network functions in salt/utils/socket_util.py
This commit moves several functions from salt/modules/network.py to
salt/utils/socket_util.py, allowing the code to be re-used to flesh out
salt/modules/win_network.py.
