+++ salt/modules/win_network.py
@@ -8,6 +8,7 @@ import re
 # Import salt libs
 import salt.utils
 from salt.utils.socket_util import sanitize_host
+
 try:
     import salt.utils.winapi
     HAS_DEPENDENCIES = True
@@ -156,95 +157,73 @@ def dig(host):
     return __salt__['cmd.run'](cmd)
 
 
-def _cidr_to_ipv4_netmask(cidr_bits):
-    '''
-    Returns an IPv4 netmask
-    '''
-    netmask = ''
-    for idx in range(4):
-        if idx:
-            netmask += '.'
-        if cidr_bits >= 8:
-            netmask += '255'
-            cidr_bits -= 8
-        else:
-            netmask += '{0:d}'.format(256 - (2 ** (8 - cidr_bits)))
-            cidr_bits = 0
-    return netmask
-
-
-def _interfaces_ipconfig(out):
-    '''
-    Returns a dictionary of interfaces with various information about each
-    (up/down state, ip address, netmask, and hwaddr)
-    '''
-    ifaces = dict()
-    iface = None
-    adapter_iface_regex = re.compile(r'adapter (\S.+):$')
-
-    for line in out.splitlines():
-        if not line:
-            continue
-        # TODO what does Windows call Infiniband and 10/40gige adapters
-        if line.startswith('Ethernet'):
-            iface = ifaces[adapter_iface_regex.search(line).group(1)]
-            iface['up'] = True
-            addr = None
-            continue
-        if iface:
-            key, val = line.split(',', 1)
-            key = key.strip(' .')
-            val = val.strip()
-            if addr and key in ('Subnet Mask'):
-                addr['netmask'] = val
-            elif key in ('IP Address', 'IPv4 Address'):
-                if 'inet' not in iface:
-                    iface['inet'] = list()
-                addr = {'address': val.rstrip('(Preferred)'),
-                        'netmask': None,
-                        'broadcast': None}  # TODO find the broadcast
-                iface['inet'].append(addr)
-            elif 'IPv6 Address' in key:
-                if 'inet6' not in iface:
-                    iface['inet'] = list()
-                # XXX What is the prefixlen!?
-                addr = {'address': val.rstrip('(Preferred)'),
-                        'prefixlen': None}
-                iface['inet6'].append(addr)
-            elif key in ('Physical Address'):
-                iface['hwaddr'] = val
-            elif key in ('Media State'):
-                # XXX seen used for tunnel adaptors
-                # might be useful
-                iface['up'] = (val != 'Media disconnected')
-
-
 def interfaces():
     '''
-    Return details about each network interface
-    '''
-    with salt.utils.winapi.Com():
-        c = wmi.WMI()
-        ifaces = {}
-        for iface in c.Win32_NetworkAdapterConfiguration(IPEnabled=1):
-            ifaces[iface.Description] = dict()
-            if iface.MACAddress:
-                ifaces[iface.Description]['hwaddr'] = iface.MACAddress
-            if iface.IPEnabled:
-                ifaces[iface.Description]['up'] = True
-                ifaces[iface.Description]['inet'] = []
-                for ip in iface.IPAddress:
-                    item = {}
-                    item['broadcast'] = ''
-                    try:
-                        item['broadcast'] = iface.DefaultIPGateway[0]
-                    except Exception:
-                        pass
-                    item['netmask'] = iface.IPSubnet[0]
-                    item['label'] = iface.Description
-                    item['address'] = ip
-                    ifaces[iface.Description]['inet'].append(item)
-            else:
-                ifaces[iface.Description]['up'] = False
-    return ifaces
+    Return a dictionary of information about all the interfaces on the minion
 
+    CLI Example::
+
+        salt '*' network.interfaces
+    '''
+    return salt.utils.socket_util.interfaces()
+
+
+def hwaddr(iface):
+    '''
+    Return the hardware address (a.k.a. MAC address) for a given interface
+
+    CLI Example::
+
+        salt '*' network.hwaddr eth0
+    '''
+    return salt.utils.socket_util.hwaddr(iface)
+
+
+def subnets():
+    '''
+    Returns a list of subnets to which the host belongs
+
+    CLI Example::
+
+        salt '*' network.subnets
+    '''
+    return salt.utils.socket_util.subnets()
+
+
+def in_subnet(cidr):
+    '''
+    Returns True if host is within specified subnet, otherwise False
+
+    CLI Example::
+
+        salt '*' network.in_subnet 10.0.0.0/16
+    '''
+    return salt.utils.socket_util.in_subnet(cidr)
+
+
+def ip_addrs(interface=None, include_loopback=False):
+    '''
+    Returns a list of IPv4 addresses assigned to the host. 127.0.0.1 is
+    ignored, unless 'include_loopback=True' is indicated. If 'interface' is
+    provided, then only IP addresses from that interface will be returned.
+
+    CLI Example::
+
+        salt '*' network.ip_addrs
+    '''
+    return salt.utils.socket_util.ip_addrs(interface=interface,
+                                           include_loopback=include_loopback)
+
+
+def ip_addrs6(interface=None, include_loopback=False):
+    '''
+    Returns a list of IPv6 addresses assigned to the host. ::1 is ignored,
+    unless 'include_loopback=True' is indicated. If 'interface' is provided,
+    then only IP addresses from that interface will be returned.
+
+    CLI Example::
+
+        salt '*' network.ip_addrs6
+    '''
+    return salt.utils.socket_util.ip_addrs6(interface=interface,
+                                            include_loopback=include_loopback)
