+++ salt/utils/socket_util.py
@@ -17,6 +17,7 @@ log = logging.getLogger(__name__)
 
 # pylint: disable-msg=C0103
 
+
 def sanitize_host(host):
     '''
     Sanitize host string.
@@ -197,7 +198,7 @@ def _interfaces_ip(out):
                             'netmask': mask,
                             'broadcast': brd,
                             'label': iflabel,
-                            })
+                        })
                         del ip_, mask, brd
                 elif type_.startswith('link'):
                     data['hwaddr'] = value
@@ -247,7 +248,7 @@ def _interfaces_ifconfig(out):
                 if mmask:
                     if mmask.group(1):
                         mmask = _number_of_set_bits_to_ipv4_netmask(
-                                int(mmask.group(1), 16))
+                            int(mmask.group(1), 16))
                     else:
                         mmask = mmask.group(2)
                     addr_obj['netmask'] = mmask
@@ -272,49 +273,252 @@ def _interfaces_ifconfig(out):
     return ret
 
 
+def linux_interfaces():
+    '''
+    Obtain interface information for *NIX/BSD variants
+    '''
+    ifaces = dict()
+    if salt.utils.which('ip'):
+        cmd1 = subprocess.Popen(
+            'ip link show',
+            shell=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.STDOUT).communicate()[0]
+        cmd2 = subprocess.Popen(
+            'ip addr show',
+            shell=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.STDOUT).communicate()[0]
+        ifaces = _interfaces_ip(cmd1 + '\n' + cmd2)
+    elif salt.utils.which('ifconfig'):
+        cmd = subprocess.Popen(
+            'ifconfig -a',
+            shell=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.STDOUT).communicate()[0]
+        ifaces = _interfaces_ifconfig(cmd)
+    return ifaces
+
+
+def _interfaces_ipconfig(out):
+    '''
+    Returns a dictionary of interfaces with various information about each
+    (up/down state, ip address, netmask, and hwaddr)
+
+    NOTE: This is not used by any function and may be able to be removed in the
+    future.
+    '''
+    ifaces = dict()
+    iface = None
+    adapter_iface_regex = re.compile(r'adapter (\S.+):$')
+
+    for line in out.splitlines():
+        if not line:
+            continue
+        # TODO what does Windows call Infiniband and 10/40gige adapters
+        if line.startswith('Ethernet'):
+            iface = ifaces[adapter_iface_regex.search(line).group(1)]
+            iface['up'] = True
+            addr = None
+            continue
+        if iface:
+            key, val = line.split(',', 1)
+            key = key.strip(' .')
+            val = val.strip()
+            if addr and key in ('Subnet Mask'):
+                addr['netmask'] = val
+            elif key in ('IP Address', 'IPv4 Address'):
+                if 'inet' not in iface:
+                    iface['inet'] = list()
+                addr = {'address': val.rstrip('(Preferred)'),
+                        'netmask': None,
+                        'broadcast': None}  # TODO find the broadcast
+                iface['inet'].append(addr)
+            elif 'IPv6 Address' in key:
+                if 'inet6' not in iface:
+                    iface['inet'] = list()
+                # XXX What is the prefixlen!?
+                addr = {'address': val.rstrip('(Preferred)'),
+                        'prefixlen': None}
+                iface['inet6'].append(addr)
+            elif key in ('Physical Address'):
+                iface['hwaddr'] = val
+            elif key in ('Media State'):
+                # XXX seen used for tunnel adaptors
+                # might be useful
+                iface['up'] = (val != 'Media disconnected')
+
+
+def win_interfaces():
+    '''
+    Obtain interface information for Windows systems
+    '''
+    with salt.utils.winapi.Com():
+        c = wmi.WMI()
+        ifaces = {}
+        for iface in c.Win32_NetworkAdapterConfiguration(IPEnabled=1):
+            ifaces[iface.Description] = dict()
+            if iface.MACAddress:
+                ifaces[iface.Description]['hwaddr'] = iface.MACAddress
+            if iface.IPEnabled:
+                ifaces[iface.Description]['up'] = True
+                ifaces[iface.Description]['inet'] = []
+                for ip in iface.IPAddress:
+                    item = {}
+                    item['broadcast'] = ''
+                    try:
+                        item['broadcast'] = iface.DefaultIPGateway[0]
+                    except Exception:
+                        pass
+                    item['netmask'] = iface.IPSubnet[0]
+                    item['label'] = iface.Description
+                    item['address'] = ip
+                    ifaces[iface.Description]['inet'].append(item)
+            else:
+                ifaces[iface.Description]['up'] = False
+    return ifaces
+
+
 def interfaces():
     '''
     Return a dictionary of information about all the interfaces on the minion
     '''
     if salt.utils.is_windows():
-        from salt.modules.win_network import interfaces as win_interfaces
-        ifaces = win_interfaces()
-        return ifaces
-
-    ifaces = dict()
-    if salt.utils.which('ip'):
-        cmd1 = subprocess.Popen(
-                'ip link show',
-                shell=True,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.STDOUT).communicate()[0]
-        cmd2 = subprocess.Popen(
-                'ip addr show',
-                shell=True,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.STDOUT).communicate()[0]
-        ifaces = _interfaces_ip(cmd1 + '\n' + cmd2)
-    elif salt.utils.which('ifconfig'):
-        cmd = subprocess.Popen(
-                'ifconfig -a',
-                shell=True,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.STDOUT).communicate()[0]
-        ifaces = _interfaces_ifconfig(cmd)
-    return ifaces
+        return win_interfaces()
+    else:
+        return linux_interfaces()
 
 
-def ip4_addrs():
+def _get_net_start(ipaddr, netmask):
+    ipaddr_octets = ipaddr.split('.')
+    netmask_octets = netmask.split('.')
+    net_start_octets = [str(int(ipaddr_octets[x]) & int(netmask_octets[x]))
+                        for x in range(0, 4)]
+    return '.'.join(net_start_octets)
+
+
+def _get_net_size(mask):
+    binary_str = ''
+    for octet in mask.split('.'):
+        binary_str += bin(int(octet))[2:].zfill(8)
+    return len(binary_str.rstrip('0'))
+
+
+def _calculate_subnet(ipaddr, netmask):
+    return '{0}/{1}'.format(_get_net_start(ipaddr, netmask),
+                            _get_net_size(netmask))
+
+
+def _ipv4_to_bits(ipaddr):
     '''
-    Return a list of ip addrs
+    Accepts an IPv4 dotted quad and returns a string representing its binary
+    counterpart
     '''
-    ret = set()
+    return ''.join([bin(int(x))[2:].rjust(8, '0') for x in ipaddr.split('.')])
+
+
+def hwaddr(iface):
+    '''
+    Return the hardware address (a.k.a. MAC address) for a given interface
+    '''
+    return interfaces().get(iface, {}).get('hwaddr', '')
+
+
+def subnets():
+    '''
+    Returns a list of subnets to which the host belongs
+    '''
     ifaces = interfaces()
-    for face in ifaces:
-        for inet in ifaces[face].get('inet', []):
-            if 'address' in inet:
-                ret.add(inet['address'])
-    return sorted(ret)
+    subnetworks = []
+
+    for ipv4_info in ifaces.values():
+        for ipv4 in ipv4_info.get('inet', []):
+            if ipv4['address'] == '127.0.0.1':
+                continue
+            network = _calculate_subnet(ipv4['address'], ipv4['netmask'])
+            subnetworks.append(network)
+    return subnetworks
+
+
+def in_subnet(cidr, addrs=None):
+    '''
+    Returns True if host is within specified subnet, otherwise False
+    '''
+    try:
+        netstart, netsize = cidr.split('/')
+        netsize = int(netsize)
+    except Exception:
+        log.error('Invalid CIDR \'{0}\''.format(cidr))
+        return False
+
+    netstart_bin = _ipv4_to_bits(netstart)
+
+    if netsize < 32 and len(netstart_bin.rstrip('0')) > netsize:
+        log.error('Invalid network starting IP \'{0}\' in CIDR '
+                  '\'{1}\''.format(netstart, cidr))
+        return False
+
+    netstart_leftbits = netstart_bin[0:netsize]
+
+    if addrs is None:
+        addrs = ip_addrs()
+
+    for ip_addr in addrs:
+        if netsize == 32:
+            if netstart == ip_addr:
+                return True
+        else:
+            ip_leftbits = _ipv4_to_bits(ip_addr)[0:netsize]
+            if netstart_leftbits == ip_leftbits:
+                return True
+    return False
+
+
+def ip_addrs(interface=None, include_loopback=False):
+    '''
+    Returns a list of IPv4 addresses assigned to the host. 127.0.0.1 is
+    ignored, unless 'include_loopback=True' is indicated. If 'interface' is
+    provided, then only IP addresses from that interface will be returned.
+    '''
+    ret = []
+    ifaces = interfaces()
+    if interface is None:
+        target_ifaces = ifaces
+    else:
+        target_ifaces = dict([(k, v) for k, v in ifaces.iteritems()
+                              if k == interface])
+        if not target_ifaces:
+            log.error('Interface {0} not found.'.format(interface))
+    for ipv4_info in target_ifaces.values():
+        for ipv4 in ipv4_info.get('inet', []):
+            if include_loopback \
+                    or (not include_loopback
+                        and ipv4['address'] != '127.0.0.1'):
+                ret.append(ipv4['address'])
+    return ret
+
+
+def ip_addrs6(interface=None, include_loopback=False):
+    '''
+    Returns a list of IPv6 addresses assigned to the host. ::1 is ignored,
+    unless 'include_loopback=True' is indicated. If 'interface' is provided,
+    then only IP addresses from that interface will be returned.
+    '''
+    ret = []
+    ifaces = interfaces()
+    if interface is None:
+        target_ifaces = ifaces
+    else:
+        target_ifaces = dict([(k, v) for k, v in ifaces.iteritems()
+                              if k == interface])
+        if not target_ifaces:
+            log.error('Interface {0} not found.'.format(interface))
+    for ipv6_info in target_ifaces.values():
+        for ipv6 in ipv6_info.get('inet6', []):
+            if include_loopback \
+                    or (not include_loopback and ipv6['address'] != '::1'):
+                ret.append(ipv6['address'])
+    return ret
 
 
 def hex2ip(hex_ip):
@@ -326,11 +530,10 @@ def hex2ip(hex_ip):
         hip = int(hex_ip, 16)
     except ValueError:
         return hex_ip
-    return '{0}.{1}.{2}.{3}'.format(
-            hip >> 24 & 255,
-            hip >> 16 & 255,
-            hip >> 8 & 255,
-            hip & 255)
+    return '{0}.{1}.{2}.{3}'.format(hip >> 24 & 255,
+                                    hip >> 16 & 255,
+                                    hip >> 8 & 255,
+                                    hip & 255)
 
 
 class IPv4Address(object):
